pyproject.toml
[project]
name = "productivity"
version = "0.1.0"
description = "Python productivity techniques"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
snowflake = [
    "snowflake-connector-python>=3.0.0",
]
db2 = [
    "ibm-db>=3.2.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]

[tool.uv]
# This project contains scripts, not an installable library.
# Tell uv not to install the project itself as a package.
package = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
